Move DefaultProfiler to a separate module
To ensure that all methods necessary to implement a custom profiler are
public.
