+++ crates/cranelift/src/compiler.rs
@@ -314,7 +314,7 @@ impl wasmtime_environ::Compiler for Compiler {
             None
         };
 
-        let timing = cranelift_codegen::timing::take_current();
+        let timing = cranelift_codegen::default_profiler::take_current();
         log::debug!("{:?} translated in {:?}", func_index, timing.total());
         log::trace!("{:?} timing info\n{}", func_index, timing);
